refactor: JS separado, return habilitado dnv
